Create utilities for dealing with clips
The previous system of downloading clips of videos isn't gonna work out, so the new plan is to either a) provide your own video to scout from, or b) select from the list of videos that are on the server. This limits what the Scouting App will be capable of when it can't connect to the server, but decreases the amount of effort needed to something I can manage. The only remaining issue with such a system would be storage restrictions, but that's something that can be dealt with later.