Fix index redirect system
The function `set_index_redirect` constructed endpoint names assuming that they were made up of the module name + the function name, but this assumption was wrong (its blueprint name + function name). Now, when `set_index_redirect` is called, it stores the function directly. Later, the endpoint is searched for when `serve` is called.