Fix sw network error
Just some faulty logic on if statements that caused fetch requests to completely fail if there was an error when `isConnected` is true. Even if there was a cached version of the page, the error would be re-thrown. This was fixed by only checking if there was a cached response for the failed request, then logging the error with `console.error`.